feat(gcf-utils): add a url for trigger source repo and include a payload hash
