+++ packages/gcf-utils/package.json
@@ -23,15 +23,16 @@
     "@google-cloud/secret-manager": "^3.0.0",
     "@google-cloud/storage": "^5.0.0",
     "@google-cloud/tasks": "^2.0.0",
+    "@octokit/plugin-enterprise-compatibility": "1.2.5",
     "@types/bunyan": "^1.8.6",
     "@types/dotenv": "^6.1.1",
     "@types/express": "^4.17.0",
     "@types/ioredis": "^4.14.8",
     "@types/lru-cache": "^5.1.0",
     "@types/sonic-boom": "^0.7.0",
-    "@octokit/plugin-enterprise-compatibility": "1.2.5",
     "express": "^4.17.1",
     "gaxios": "^3.0.0",
+    "md5": "^2.2.1",
     "pino": "^6.3.2",
     "probot": "^9.11.2",
     "tmp": "^0.2.0",
@@ -44,6 +45,7 @@
     "@types/sinon": "^9.0.0",
     "@types/tmp": "^0.2.0",
     "@types/yargs": "^15.0.0",
+    "@types/md5": "^2.2.0",
     "c8": "^7.1.0",
     "cross-env": "^7.0.0",
     "dotenv": "^8.0.0",
