+++ packages/gcf-utils/src/logging/logging-octokit-plugin.ts
@@ -14,7 +14,7 @@
 //
 // eslint-disable-next-line node/no-extraneous-import
 import {Octokit} from '@octokit/rest';
-import {logger} from './gcf-utils';
+import {logger} from '../gcf-utils';
 import {
   OctokitRequestParser as parser,
   GitHubActionDetails,
